Servo moteur fonctionnel

+++ Drivers/Include/MyTimer.h
@@ -45,7 +45,7 @@ void MyTimer_PWM_set_CCR( TIM_TypeDef * Timer, int Channel, int8_t CCR);
 * @param : - TIM_TypeDef * Timer : Timer concerné
 * 				 - int8_t ratio :
 ***************************************************************************************************/
-int8_t MyTimer_PWM_calculer_CCR(TIM_TypeDef * Timer, int8_t ratio);
+int8_t MyTimer_PWM_calculer_CCR(TIM_TypeDef * Timer, int ratio);
 
 
 /**/
